+++ src/classes/ExtendedClient.ts
@@ -5,6 +5,7 @@ import Modal from './Modal';
 import Timer from './Timer';
 import Logger from './Logger';
 import { Ban, PrismaClient } from '@prisma/client';
+import StringSelectMenu from './StringSelectMenu';
 
 export interface ClientCache {
   bans: Collection<number, Ban>;
@@ -15,6 +16,7 @@ export default class ExtendedClient extends Client {
   public commands: Collection<string, Command>;
   public buttons: Collection<string, Button>;
   public modals: Collection<string, Modal>;
+  public stringSelectMenus: Collection<string, StringSelectMenu>;
   public timers: Collection<string, Timer>;
   public prisma: PrismaClient;
   public color: ColorResolvable;
@@ -27,6 +29,7 @@ export default class ExtendedClient extends Client {
     this.commands = new Collection();
     this.buttons = new Collection();
     this.modals = new Collection();
+    this.stringSelectMenus = new Collection();
     this.timers = new Collection();
     this.prisma = new PrismaClient();
     this.color = '#5865F2';
