+++ src/exceptions/StringSelectMenuNotFoundException.ts
@@ -1,0 +1,8 @@
+import Exception from './Exception';
+
+export default class StringSelectMenuNotFoundException extends Exception {
+  constructor(command: string) {
+    super(`Could not find string select menu ${command}`);
+    this.name = 'StringSelectMenuNotFoundException';
+  }
+}
